FEAT: Updated hugging face url chat to use httpx
